Backend Init
Setup a skeleton for how the backend will work, utilizing Node.js and Express. The current API endpoints are for creating a game, submitting a turn, and querying the game state. The endpoints aren't fully functional, but serve as a proof of concept. I also created a docs directory at the root in addition to the frontend and backend dirs to document the backend progress for the frontend.